Feat : 등록및조회 테스트

+++ beer_community/src/main/java/team_project/beer_community/dto/BeerDetailDto.java
@@ -11,9 +11,9 @@ public class BeerDetailDto {
     private Long id;
     private String imageUrl;
     private String beerName;
-    private float totalPoint;
+    private double totalPoint;
     private String information;
-    private float alcoholDegree;
+    private double alcoholDegree;
     private List<Taste> tastes;
 
     public BeerDetailDto(Beer beer, List<Taste> tastes) {
